confidence as float

+++ DataLayer/Entities/RecognitionResult.cs
@@ -6,7 +6,7 @@ namespace DataLayer.Entities
     public class RecognitionResult : EntityBase
     {
         public int IdentifiedPersonId { get; set; }
-        public int Confidence { get; set; }
+        public float Confidence { get; set; }
         public int NeuralNetworkFileId { get; set; }
         public NeuralNetworkFile NeuralNetworkFile { get; set; }
         public int RecognitionId { get; set; }
